Add files via upload
These three are the real template. To make use of them will require some editing. I spent way too much time trying to figure out how to do this in one file/page. Using this method, we can create a set of files where each "landingtwo" can redirect back if wrong or forward to the next "landing", assuming we still want them to do the activities in a specific order. 

The page I found this on is https://webdevtrick.com/simple-quiz-in-php-source-code/.